requirements.txt
absl-py==2.1.0
albucore==0.0.13
albumentations==1.4.13
annotated-types==0.7.0
anyio==4.6.0
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
contourpy==1.2.1
cssselect2==0.7.0
cycler==0.12.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
einops==0.8.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
fastprogress==1.0.3
filelock==3.14.0
Flask==3.0.3
fonttools==4.53.0
fsspec==2025.2.0
gensim==4.3.2
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.23.3
idna==3.7
imageio==2.34.1
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
kiwisolver==1.4.5
langdetect==1.0.9
lazy_loader==0.4
Levenshtein==0.25.1
llvmlite==0.43.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
mplcursors==0.5.3
mpmath==1.3.0
multiprocess==0.70.16
networkx==3.3
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
openai==1.50.1
opencv-python==4.10.0.82
opencv-python-headless==4.10.0.82
OpenNMT-py==2.2.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
psutil==6.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydyf==0.10.0
pylena==0.1.3
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyonmttok==1.37.1
pyparsing==3.1.2
pypdfium2==4.30.0
pyphen==0.15.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-doctr==0.8.1
pytz==2024.1
PyYAML==6.0.1
qudida==0.0.4
rapidfuzz==3.9.3
rdkit-pypi==2023.3.1b1
redun==0.25.0
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-image==0.23.2
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.2.0
shapely==2.0.4
six==1.16.0
smart-open==7.0.4
SmilesPE==0.0.3
sniffio==1.3.1
soupsieve==2.6
sympy==1.12.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tifffile==2024.5.22
timeout-decorator==0.5.0
timm==0.4.12
tinycss2==1.3.0
tokenizers==0.19.1
tomli==2.0.1
torch==1.13.1
torchtext==0.5.0
torchvision==0.14.1
tqdm==4.66.4
transformers==4.41.2
triton==2.3.1
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.74
ultralytics-thop==2.0.0
Unidecode==1.3.8
urllib3==2.2.1
waitress==3.0.0
weasyprint==62.3
webencodings==0.5.1
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.19.2
zopfli==0.2.3
